refactor: more separation of logic from utils.py
