+++ src/ExcelWriter.py
@@ -6,10 +6,10 @@ from tqdm import tqdm
 
 from tornado.gen import sleep
 
-from Utils.utils import (
+from Utils.output_utils import cell_formatting
+from Utils.metrics_utils import (
     get_percentage_value,
     get_metrics,
-    cell_formatting
 )
