+++ src/model/EvaluationSummary.py
@@ -1,4 +1,4 @@
-from Utils.utils import (
+from Utils.metrics_utils import (
     count_actual_values,
     count_predicted_values,
     calculate_confusion_matrix_metrics,
